chore: remove FederatedIngress alpha feature
