+++ charts/kubefed/charts/controllermanager/templates/kubefedconfig.yaml
@@ -26,8 +26,6 @@ spec:
     configuration: {{ .Values.featureGates.PushReconciler | default "Enabled" | quote }}
   - name: SchedulerPreferences
     configuration: {{ .Values.featureGates.SchedulerPreferences | default "Enabled" | quote }}
-  - name: FederatedIngress
-    configuration: {{ .Values.featureGates.FederatedIngress | default "Disabled" | quote }}
   # NOTE: Commented feature gate to fix https://github.com/kubernetes-sigs/kubefed/issues/1333
   #- name: RawResourceStatusCollection
   #  configuration: {{ .Values.featureGates.RawResourceStatusCollection | default "Disabled" | quote }}
