+++ charts/kubefed/values.yaml
@@ -23,7 +23,6 @@ controllermanager:
   featureGates:
     PushReconciler:
     SchedulerPreferences:
-    FederatedIngress:
     RawResourceStatusCollection:
 
   ## common node selector
