+++ config/kubefedconfig.yaml
@@ -20,8 +20,6 @@ spec:
     configuration: "Enabled"
   - name: SchedulerPreferences
     configuration: "Enabled"
-  - name: FederatedIngress
-    configuration: "Enabled"
   clusterHealthCheck:
     failureThreshold: 3
     period: 10s
